+++ src/iomusxml.cpp
@@ -2103,6 +2103,7 @@ void MusicXmlInput::ReadMusicXmlDirection(
     if (containsWords && !containsTempo && !containsDynamics) {
         pugi::xpath_node_set words = node.select_nodes("direction-type/*[self::words or self::coda or self::segno]");
         defaultY = words.first().node().attribute("default-y").as_int();
+        defaultY = (defaultY * 10) + words.first().node().attribute("relative-y").as_int();
         std::string wordStr = words.first().node().text().as_string();
         if (wordStr.rfind("cresc", 0) == 0 || wordStr.rfind("dim", 0) == 0 || wordStr.rfind("decresc", 0) == 0) {
             containsDynamics = true;
@@ -2128,7 +2129,7 @@ void MusicXmlInput::ReadMusicXmlDirection(
             }
 
             TextRendition(words, dir);
-            defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 200;
+            defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 2000;
             dir->SetVgrp(defaultY);
             m_controlElements.push_back({ measureNum, dir });
             m_dirStack.push_back(dir);
@@ -2183,9 +2184,12 @@ void MusicXmlInput::ReadMusicXmlDirection(
         }
 
         TextRendition(dynamics, dynam);
-        if (defaultY == 0) defaultY = dynamics.first().node().attribute("default-y").as_int();
+        if (defaultY == 0) {
+            defaultY = dynamics.first().node().attribute("default-y").as_int();
+            defaultY = (defaultY * 10) + dynamics.first().node().attribute("relative-y").as_int();
+        }
         // parse the default_y attribute and transform to vgrp value, to vertically align dynamics and directives
-        defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 200;
+        defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 2000;
         dynam->SetVgrp(defaultY);
         m_controlElements.push_back({ measureNum, dynam });
         m_dynamStack.push_back(dynam);
@@ -2285,8 +2289,9 @@ void MusicXmlInput::ReadMusicXmlDirection(
                 hairpin->SetStaff(hairpin->AttStaffIdent::StrToXsdPositiveIntegerList(std::to_string(1 + staffOffset)));
             }
             int defaultY = wedge->node().attribute("default-y").as_int();
+            defaultY = (defaultY * 10) + wedge->node().attribute("relative-y").as_int();
             // parse the default_y attribute and transform to vgrp value, to vertically align hairpins
-            defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 200;
+            defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 2000;
             hairpin->SetVgrp(defaultY);
             // match new hairpin to existing hairpin stop
             for (auto iter = m_hairpinStopStack.begin(); iter != m_hairpinStopStack.end(); ++iter) {
@@ -2364,13 +2369,13 @@ void MusicXmlInput::ReadMusicXmlDirection(
             pedal->SetDir(ConvertPedalTypeToDir(pedalType));
             if (pedalLine) pedal->SetForm(PEDALSTYLE_line);
             if (xmlPedal.attribute("abbreviated")) {
-                AttModule::SetExternalsymbols(pedal, "glyph.auth", "smufl");
+                pedal->SetGlyphAuth("smufl");
                 AttModule::SetExternalsymbols(pedal, "glyph.num", "U+E651");
             }
             if (pedalType == "sostenuto") {
                 pedal->SetFunc("sostenuto");
                 if (xmlPedal.attribute("abbreviated")) {
-                    AttModule::SetExternalsymbols(pedal, "glyph.auth", "smufl");
+                    pedal->SetGlyphAuth("smufl");
                     AttModule::SetExternalsymbols(pedal, "glyph.num", "U+E65A");
                 }
             }
@@ -2388,8 +2393,9 @@ void MusicXmlInput::ReadMusicXmlDirection(
             pedal->SetTstamp(timeStamp);
             if (pedalType == "stop") pedal->SetTstamp(timeStamp - 0.1);
             int defaultY = xmlPedal.attribute("default-y").as_int();
+            defaultY = (defaultY * 10) + xmlPedal.attribute("relative-y").as_int();
             // parse the default_y attribute and transform to vgrp value, to vertically align pedal starts and stops
-            defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 200;
+            defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 2000;
             pedal->SetVgrp(defaultY);
             m_controlElements.push_back({ measureNum, pedal });
             m_pedalStack.push_back(pedal);
@@ -2696,11 +2702,11 @@ void MusicXmlInput::ReadMusicXmlNote(
                 AddLayerElement(layer, bTrem);
                 m_elementStackMap.at(layer).push_back(bTrem);
                 if (HasAttributeWithValue(tremolo.node(), "type", "unmeasured")) {
-                    bTrem->SetForm(bTremLog_FORM_unmeas);
+                    bTrem->SetForm(tremForm_FORM_unmeas);
                     tremSlashNum = 0;
                 }
                 else {
-                    bTrem->SetForm(bTremLog_FORM_meas);
+                    bTrem->SetForm(tremForm_FORM_meas);
                 }
             }
         }
@@ -2870,6 +2876,13 @@ void MusicXmlInput::ReadMusicXmlNote(
                 note->SetOct(octaveNum);
             }
         }
+        else if (node.child("unpitched")) {
+            pugi::xml_node unpitched = node.child("unpitched");
+            const std::string stepStr = unpitched.child("display-step").text().as_string();
+            const int octaveNum = unpitched.child("display-octave").text().as_int();
+            const int loc = note->CalcLoc(ConvertStepToPitchName(stepStr), octaveNum, -2);
+            note->SetLoc(loc);
+        }
 
         // dynamics (MIDI velocity)
         const float dynamics = node.attribute("dynamics").as_float(-1.0);
@@ -2883,6 +2896,7 @@ void MusicXmlInput::ReadMusicXmlNote(
             note->SetHeadColor(notehead.attribute("color").as_string());
             note->SetHeadShape(ConvertNotehead(notehead.text().as_string()));
             if (notehead.attribute("parentheses").as_bool()) note->SetHeadMod(NOTEHEADMODIFIER_paren);
+            note->SetGlyphName(notehead.attribute("smufl").as_string());
             auto noteHeadFill = notehead.attribute("filled");
             if (noteHeadFill) note->SetHeadFill(noteHeadFill.as_bool() ? FILL_solid : FILL_void);
             if (!std::strncmp(notehead.text().as_string(), "none", 4)) note->SetHeadVisible(BOOLEAN_false);
@@ -2915,6 +2929,7 @@ void MusicXmlInput::ReadMusicXmlNote(
                 chord->SetDurPpq(duration);
                 if (dots > 0) chord->SetDots(dots);
                 chord->SetStemDir(stemDir);
+                if (!strcmp(notehead.text().as_string(), "cluster")) chord->SetCluster(CLUSTER_white);
                 if (stemText == "none") chord->SetStemVisible(BOOLEAN_false);
                 if (tremSlashNum > 0) {
                     chord->SetStemMod(chord->AttStems::StrToStemmodifier(std::to_string(tremSlashNum) + "slash"));
@@ -3121,7 +3136,7 @@ void MusicXmlInput::ReadMusicXmlNote(
                 // color
                 meiSlur->SetColor(slur.attribute("color").as_string());
                 // lineform
-                meiSlur->SetLform(meiSlur->AttCurveRend::StrToLineform(slur.attribute("line-type").as_string()));
+                meiSlur->SetLform(meiSlur->AttLineRendBase::StrToLineform(slur.attribute("line-type").as_string()));
                 if (slur.attribute("id")) meiSlur->SetID(slur.attribute("id").as_string());
                 meiSlur->SetStartid("#" + note->GetID());
                 // add it to the stack
@@ -3288,7 +3303,8 @@ void MusicXmlInput::ReadMusicXmlNote(
         // place
         dynam->SetPlace(dynam->AttPlacementRelStaff::StrToStaffrel(xmlDynam.attribute("placement").as_string()));
         int defaultY = xmlDynam.attribute("default-y").as_int();
-        defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 200;
+        defaultY = (defaultY * 10) + xmlDynam.attribute("relative-y").as_int();
+        defaultY = (defaultY < 0) ? std::abs(defaultY) : defaultY + 2000;
         dynam->SetVgrp(defaultY);
         std::string dynamStr;
         for (pugi::xml_node xmlDynamPart : xmlDynam.children()) {
@@ -3676,14 +3692,14 @@ void MusicXmlInput::ReadMusicXmlPrint(pugi::xml_node node, Section *section)
     assert(node);
     assert(section);
 
-    if (HasAttributeWithValue(node, "new-system", "yes")) {
-        Sb *sb = new Sb();
-        section->AddChild(sb);
-    }
-
-    if (HasAttributeWithValue(node, "new-page", "yes")) {
+    if (node.attribute("new-page").as_bool()) {
         Pb *pb = new Pb();
         section->AddChild(pb);
+    }
+
+    if (node.attribute("new-system").as_bool()) {
+        Sb *sb = new Sb();
+        section->AddChild(sb);
     }
 }
 
@@ -3828,7 +3844,7 @@ void MusicXmlInput::ReadMusicXmlTies(
             tie->SetColor(xmlTie.attribute("color").as_string());
             // placement and orientation
             tie->SetCurvedir(InferCurvedir(xmlTie));
-            tie->SetLform(tie->AttCurveRend::StrToLineform(xmlTie.attribute("line-type").as_string()));
+            tie->SetLform(tie->AttLineRendBase::StrToLineform(xmlTie.attribute("line-type").as_string()));
             if (xmlTie.attribute("id")) tie->SetID(xmlTie.attribute("id").as_string());
             // add it to the stack
             m_controlElements.push_back({ measureNum, tie });
@@ -3841,7 +3857,7 @@ void MusicXmlInput::ReadMusicXmlTies(
             lv->SetColor(xmlTie.attribute("color").as_string());
             // placement and orientation
             lv->SetCurvedir(InferCurvedir(xmlTie));
-            lv->SetLform(lv->AttCurveRend::StrToLineform(xmlTie.attribute("line-type").as_string()));
+            lv->SetLform(lv->AttLineRendBase::StrToLineform(xmlTie.attribute("line-type").as_string()));
             if (xmlTie.attribute("id")) lv->SetID(xmlTie.attribute("id").as_string());
             // add it to the stack
             m_controlElements.push_back({ measureNum, lv });
@@ -3934,7 +3950,7 @@ KeySig *MusicXmlInput::ConvertKey(const pugi::xml_node &key)
         if (key.child("mode")) {
             const std::string xmlMode = key.child("mode").text().as_string();
             if (std::strncmp(xmlMode.c_str(), "none", 4)) {
-                keySig->SetMode(keySig->AttKeySigLog::StrToMode(xmlMode));
+                keySig->SetMode(keySig->AttKeyMode::StrToMode(xmlMode));
             }
         }
     }
@@ -3948,6 +3964,10 @@ KeySig *MusicXmlInput::ConvertKey(const pugi::xml_node &key)
                 if (std::strncmp(keyStep.next_sibling().next_sibling().name(), "key-accidental", 14) == 0) {
                     keyAccid->SetAccid(
                         ConvertAccidentalToAccid(keyStep.next_sibling().next_sibling().text().as_string()));
+                    keyAccid->SetGlyphName(keyStep.next_sibling().next_sibling().attribute("smufl").as_string());
+                }
+                else if (!keyAccid->HasAccid()) {
+                    LogWarning("MusicXML import: Could not properly set keyAccid");
                 }
             }
             keySig->AddChild(keyAccid);
@@ -4004,7 +4024,7 @@ curvature_CURVEDIR MusicXmlInput::CombineCurvedir(curvature_CURVEDIR startDir, c
 
 data_ACCIDENTAL_WRITTEN MusicXmlInput::ConvertAccidentalToAccid(const std::string &value)
 {
-    static const std::map<std::string, data_ACCIDENTAL_WRITTEN> Accidental2Accid{
+    static const std::map<std::string, data_ACCIDENTAL_WRITTEN> Accidental2Accid{ //
         { "sharp", ACCIDENTAL_WRITTEN_s }, //
         { "natural", ACCIDENTAL_WRITTEN_n }, //
         { "flat", ACCIDENTAL_WRITTEN_f }, //
@@ -4023,12 +4043,19 @@ data_ACCIDENTAL_WRITTEN MusicXmlInput::ConvertAccidentalToAccid(const std::strin
         { "natural-up", ACCIDENTAL_WRITTEN_nu }, //
         { "flat-down", ACCIDENTAL_WRITTEN_fd }, //
         { "flat-up", ACCIDENTAL_WRITTEN_fu }, //
+        { "double-sharp-down", ACCIDENTAL_WRITTEN_xd }, //
+        { "double-sharp-up", ACCIDENTAL_WRITTEN_xu }, //
+        { "flat-flat-down", ACCIDENTAL_WRITTEN_ffd }, //
+        { "flat-flat-up", ACCIDENTAL_WRITTEN_ffu }, //
         { "triple-sharp", ACCIDENTAL_WRITTEN_ts }, //
         { "triple-flat", ACCIDENTAL_WRITTEN_tf }, //
         { "slash-quarter-sharp", ACCIDENTAL_WRITTEN_bms }, //
         { "slash-sharp", ACCIDENTAL_WRITTEN_ks }, //
         { "slash-flat", ACCIDENTAL_WRITTEN_bf }, //
-        { "double-slash-flat", ACCIDENTAL_WRITTEN_bmf } //
+        { "double-slash-flat", ACCIDENTAL_WRITTEN_bmf }, //
+        { "sori", ACCIDENTAL_WRITTEN_sori }, //
+        { "koron", ACCIDENTAL_WRITTEN_koron }, //
+        { "other", ACCIDENTAL_WRITTEN_NONE }
     };
 
     const auto result = Accidental2Accid.find(value);
@@ -4044,6 +4071,7 @@ data_ACCIDENTAL_GESTURAL MusicXmlInput::ConvertAlterToAccid(const float value)
 {
     static const std::map<float, data_ACCIDENTAL_GESTURAL> Alter2Accid{
         { -3, ACCIDENTAL_GESTURAL_tf }, //
+        { -2.5, ACCIDENTAL_GESTURAL_ffd }, //
         { -2, ACCIDENTAL_GESTURAL_ff }, //
         { -1.5, ACCIDENTAL_GESTURAL_fd }, //
         { -1, ACCIDENTAL_GESTURAL_f }, //
@@ -4053,6 +4081,7 @@ data_ACCIDENTAL_GESTURAL MusicXmlInput::ConvertAlterToAccid(const float value)
         { 1, ACCIDENTAL_GESTURAL_s }, //
         { 1.5, ACCIDENTAL_GESTURAL_su }, //
         { 2, ACCIDENTAL_GESTURAL_ss }, //
+        { 2.5, ACCIDENTAL_GESTURAL_xu }, //
         { 2, ACCIDENTAL_GESTURAL_ts } //
     };
 
@@ -4061,7 +4090,6 @@ data_ACCIDENTAL_GESTURAL MusicXmlInput::ConvertAlterToAccid(const float value)
         return result->second;
     }
 
-    LogWarning("MusicXML import: Unsupported alter value '%.1f'", value);
     return ACCIDENTAL_GESTURAL_NONE;
 }