+++ src/view_page.cpp
@@ -135,9 +135,9 @@ void View::SetScoreDefDrawingWidth(DeviceContext *dc, ScoreDef *scoreDef)
     }
 
     // longest key signature of the staffDefs
-    const ListOfObjects &scoreDefList = scoreDef->GetList(scoreDef); // make sure it's initialized
-    for (ListOfObjects::const_iterator it = scoreDefList.begin(); it != scoreDefList.end(); ++it) {
-        StaffDef *staffDef = vrv_cast<StaffDef *>(*it);
+    const ListOfObjects &childList = scoreDef->GetList(); // make sure it's initialized
+    for (Object *child : childList) {
+        StaffDef *staffDef = vrv_cast<StaffDef *>(child);
         assert(staffDef);
         if (!staffDef->HasKeySigInfo()) continue;
         KeySig *keySig = staffDef->GetKeySig();
@@ -541,8 +541,8 @@ void View::DrawLabels(
     LabelAbbr *labelAbbr = vrv_cast<LabelAbbr *>(object->FindDescendantByType(LABELABBR, 1));
     Object *graphic = label;
 
-    std::u32string labelStr = (label) ? label->GetText(label) : U"";
-    std::u32string labelAbbrStr = (labelAbbr) ? labelAbbr->GetText(labelAbbr) : U"";
+    std::u32string labelStr = label ? label->GetText() : U"";
+    std::u32string labelAbbrStr = labelAbbr ? labelAbbr->GetText() : U"";
 
     if (abbreviations) {
         labelStr = labelAbbrStr;
@@ -586,7 +586,7 @@ void View::DrawLabels(
     if (labelAbbr && !abbreviations && (labelAbbrStr.length() > 0)) {
         TextExtend extend;
         std::vector<std::u32string> lines;
-        labelAbbr->GetTextLines(labelAbbr, lines);
+        labelAbbr->GetTextLines(lines);
         int maxLength = 0;
         for (std::u32string &line : lines) {
             dc->GetTextExtent(line, &extend, true);
@@ -1116,7 +1116,7 @@ void View::DrawMeterSigGrp(DeviceContext *dc, Layer *layer, Staff *staff)
     assert(staff);
 
     MeterSigGrp *meterSigGrp = layer->GetStaffDefMeterSigGrp();
-    ListOfObjects childList = meterSigGrp->GetList(meterSigGrp);
+    ListOfObjects childList = meterSigGrp->GetList();
 
     // Ignore invisible meter signatures and those without count
     childList.erase(std::remove_if(childList.begin(), childList.end(),
@@ -1246,7 +1246,9 @@ void View::DrawStaff(DeviceContext *dc, Staff *staff, Measure *measure, System *
         this->DrawStaffLines(dc, staff, measure, system);
     }
 
-    this->DrawStaffDef(dc, staff, measure);
+    if (staffDef && (staffDef->GetNotationtype() != NOTATIONTYPE_neume) && (m_doc->GetType() != Facs)) {
+        this->DrawStaffDef(dc, staff, measure);
+    }
 
     if (!staff->GetLedgerLinesAbove().empty()) {
         this->DrawLedgerLines(dc, staff, staff->GetLedgerLinesAbove(), false, false);
@@ -1973,7 +1975,7 @@ void View::DrawAnnot(DeviceContext *dc, EditorialElement *element, bool isTextEl
 
     Annot *annot = vrv_cast<Annot *>(element);
     assert(annot);
-    dc->AddDescription(UTF32to8(annot->GetText(annot)));
+    dc->AddDescription(UTF32to8(annot->GetText()));
 
     if (isTextElement) {
         dc->EndTextGraphic(element, this);